+++ package/src/server/components/server_provider.tsx
@@ -188,6 +188,7 @@ export default async function LocationzationProvider({ language, messages, stati
         initialAuthUser={initialAuthUser}
         skipAuthProvider={!autoWireClientProvider}
         analyticsConfig={analyticsConfig}
+        autoAnalyticsEventsConfig={config.cookieConsent?.autoAnalyticsEvents}
         requiresConsent={requiresConsent}
         autoWireDialogs={config.cookieConsent?.autoWireDialogs !== false}
         dialogProps={config.cookieConsent?.dialogProps}