+++ package/src/server/components/server_provider_static.tsx
@@ -126,6 +126,7 @@ export default async function LocationzationProvider({ language, messages, child
         language={language}
         messages={messagesValue}
         analyticsConfig={analyticsConfig}
+        autoAnalyticsEventsConfig={config.cookieConsent?.autoAnalyticsEvents}
         requiresConsent={requiresConsent}
         autoWireDialogs={config.cookieConsent?.autoWireDialogs !== false}
         dialogProps={config.cookieConsent?.dialogProps}